intro_cdkpipelines/requirements.txt
-e .
attrs==19.3.0
aws-cdk.assets==1.55.0
aws-cdk.aws-apigateway==1.55.0
aws-cdk.aws-applicationautoscaling==1.55.0
aws-cdk.aws-autoscaling==1.55.0
aws-cdk.aws-autoscaling-common==1.55.0
aws-cdk.aws-autoscaling-hooktargets==1.55.0
aws-cdk.aws-batch==1.55.0
aws-cdk.aws-certificatemanager==1.55.0
aws-cdk.aws-cloudformation==1.55.0
aws-cdk.aws-cloudfront==1.55.0
aws-cdk.aws-cloudwatch==1.55.0
aws-cdk.aws-codebuild==1.55.0
aws-cdk.aws-codecommit==1.55.0
aws-cdk.aws-codedeploy==1.55.0
aws-cdk.aws-codeguruprofiler==1.55.0
aws-cdk.aws-codepipeline==1.55.0
aws-cdk.aws-codepipeline-actions==1.55.0
aws-cdk.aws-cognito==1.55.0
aws-cdk.aws-ec2==1.55.0
aws-cdk.aws-ecr==1.55.0
aws-cdk.aws-ecr-assets==1.55.0
aws-cdk.aws-ecs==1.55.0
aws-cdk.aws-efs==1.55.0
aws-cdk.aws-elasticloadbalancing==1.55.0
aws-cdk.aws-elasticloadbalancingv2==1.55.0
aws-cdk.aws-events==1.55.0
aws-cdk.aws-events-targets==1.55.0
aws-cdk.aws-iam==1.55.0
aws-cdk.aws-kinesis==1.55.0
aws-cdk.aws-kms==1.55.0
aws-cdk.aws-lambda==1.55.0
aws-cdk.aws-logs==1.55.0
aws-cdk.aws-route53==1.55.0
aws-cdk.aws-route53-targets==1.55.0
aws-cdk.aws-s3==1.55.0
aws-cdk.aws-s3-assets==1.55.0
aws-cdk.aws-sam==1.55.0
aws-cdk.aws-secretsmanager==1.55.0
aws-cdk.aws-servicecatalog==1.55.0
aws-cdk.aws-servicediscovery==1.55.0
aws-cdk.aws-sns==1.55.0
aws-cdk.aws-sns-subscriptions==1.55.0
aws-cdk.aws-sqs==1.55.0
aws-cdk.aws-ssm==1.55.0
aws-cdk.aws-stepfunctions==1.55.0
aws-cdk.cloud-assembly-schema==1.55.0
aws-cdk.core==1.55.0
aws-cdk.custom-resources==1.55.0
aws-cdk.cx-api==1.55.0
aws-cdk.pipelines==1.55.0
aws-cdk.region-info==1.55.0
cattrs==1.0.0
constructs==3.0.4
jsii==1.9.0
publication==0.0.3
python-dateutil==2.8.1
six==1.15.0
typing-extensions==3.7.4.2
